+++ example/lib/demos/svg_icons/svg_icons_scene.dart
@@ -2,7 +2,7 @@ import 'package:graphx/graphx.dart';
 import 'test_icons.dart';
 // import 'test_svg_scene.dart';
 
-class SvgIconsScene extends Sprite {
+class SvgIconsScene extends GSprite {
   @override
   void addedToStage() {
     addChild(TestIcons());