+++ lib/src/utils/utils.dart
@@ -4,6 +4,7 @@ export 'context_utils.dart';
 export 'http/http.dart';
 export 'interfases.dart';
 export 'keyboard_util.dart';
+export 'layout_utils.dart';
 export 'list_utils.dart';
 export 'math_utils.dart';
 export 'matrix_utils.dart';